refactor: separate plist handling to file
